api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dateutil==2.8.2

